+++ app/services/photoprism/request_photos.rb
@@ -13,7 +13,7 @@ class Photoprism::RequestPhotos
     @user = user
     @photoprism_api_base_url = "#{user.safe_settings.photoprism_url}/api/v1/photos"
     @photoprism_api_key = user.safe_settings.photoprism_api_key
-    @start_date = start_date
+    @start_date = start_date || '1970-01-01'
     @end_date = end_date
   end
 
@@ -102,11 +102,10 @@ class Photoprism::RequestPhotos
   end
 
   def time_framed_data(data, start_date, end_date)
+    range_start = start_date.to_datetime
+    range_end = (end_date || Time.current).to_date.end_of_day.to_datetime
     data.flatten.select do |photo|
-      taken_at = DateTime.parse(photo['TakenAtLocal'])
-      end_date ||= Time.current
-
-      taken_at.between?(start_date.to_datetime, end_date.to_datetime)
+      DateTime.parse(photo['TakenAtLocal']).between?(range_start, range_end)
     end
   end